web/myapp/ai_processor.py
from transformers import BertTokenizer, BertModel
import torch
import os
from google.cloud import aiplatform
from dotenv import load_dotenv
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Load environment variables
load_dotenv()

class DreamAnalyzer:
    def __init__(self):
        # Initialize BERT
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
        self.bert_model = BertModel.from_pretrained('bert-base-uncased')
        
        # Initialize Gemini client
        self.gemini_key = os.getenv('GEMINI_API_KEY')
        if not self.gemini_key:
            raise ValueError("GEMINI_API_KEY not found in environment variables")

    async def extract_themes(self, dream_text):
        """Extract themes using BERT model"""
        try:
            # Tokenize and prepare input
            inputs = self.tokenizer(
                dream_text,
                return_tensors="pt",
                max_length=512,
                truncation=True,
                padding=True
            )
            
            # Get BERT embeddings
            with torch.no_grad():
                outputs = self.bert_model(**inputs)
            
            # Process embeddings to extract themes
            # This is a placeholder - implement your theme extraction logic here
            themes = ["placeholder_theme_1", "placeholder_theme_2"]
            
            return themes
        except Exception as e:
            logger.error(f"Error in BERT processing: {str(e)}")
            raise

    async def get_gemini_analysis(self, dream_text, themes):
        """Get detailed analysis from Gemini API"""
        try:
            # Construct prompt with dream and themes
            prompt = f"""Analyze this dream and its themes:
            Dream: {dream_text}
            Extracted Themes: {', '.join(themes)}
            
            Provide a detailed psychological analysis including:
            1. Key symbols and their meanings
            2. Emotional undertones
            3. Possible interpretations
            4. Connections to the dreamer's psyche"""
            
            # TODO: Implement actual Gemini API call
            # This is a placeholder - implement your Gemini API integration
            analysis = "Placeholder Gemini analysis"
            
            return analysis
        except Exception as e:
            logger.error(f"Error in Gemini processing: {str(e)}")
            raise

    async def analyze_dream(self, dream_text):
        """Complete dream analysis pipeline"""
        try:
            # Extract themes using BERT
            themes = await self.extract_themes(dream_text)
            
            # Get detailed analysis from Gemini
            analysis = await self.get_gemini_analysis(dream_text, themes)
            
            return {
                "themes": themes,
                "analysis": analysis,
                "dream_text": dream_text
            }
        except Exception as e:
            logger.error(f"Error in dream analysis: {str(e)}")
            raise 