requirements.txt
transformers==4.35.2
torch==2.1.1
google-cloud-aiplatform==1.36.4
python-dotenv==1.0.0
django-rest-framework==3.14.0
supabase==2.0.3 